Use atoi(3) for int
timerequest and channel are of type int. We must use atoi(3) instead of
atol(3) to convert them.

Problem reported on Mac OS X.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/HandlerTest@7108 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ Host/handler_test.c
@@ -161,7 +161,7 @@ int main(int argc, char *argv[])
 				break;
 
 			case 't':
-				timerequest = atol(optarg);
+				timerequest = atoi(optarg);
 				printf("time request: %d\n", timerequest);
 				break;
 
@@ -191,7 +191,7 @@ int main(int argc, char *argv[])
 				break;
 
 			case 'c':
-				channel = atol(optarg);
+				channel = atoi(optarg);
 				printf("Usinf channel: %d\n", channel);
 				break;
